src/engagerunner/platforms/youtube.py
"""YouTube platform implementation."""

import logging

from browser_use import Agent
from langchain_anthropic import ChatAnthropic

from engagerunner.models import Comment, Platform
from engagerunner.platforms.base import BasePlatform

logger = logging.getLogger(__name__)


class YouTubePlatform(BasePlatform):
    """YouTube-specific platform implementation."""

    platform = Platform.YOUTUBE

    async def navigate_to_url(self, url: str) -> None:
        """Navigate to a YouTube video URL.

        Args:
            url: YouTube video URL
        """
        logger.info("Navigating to %s", url)
        agent = Agent(  # type: ignore[var-annotated]
            task=f"Navigate to {url} and wait for the page to load completely",
            llm=ChatAnthropic(model="claude-sonnet-4-20250514"),
            browser=self.browser,
        )
        await agent.run()

    async def read_comments(self, url: str, max_comments: int = 10) -> list[Comment]:
        """Read comments from a YouTube video.

        Args:
            url: YouTube video URL
            max_comments: Maximum number of comments to retrieve

        Returns:
            List of Comment objects
        """
        logger.info("Reading up to %s comments from %s", max_comments, url)

        # Navigate to video
        await self.navigate_to_url(url)

        # Use browser-use agent to scroll and extract comments
        agent = Agent(  # type: ignore[var-annotated]
            task=(
                f"Scroll down to load the comments section. "
                f"Extract the first {max_comments} comments with their author names, "
                f"comment text, and timestamps. Return the data in JSON format."
            ),
            llm=ChatAnthropic(model="claude-sonnet-4-20250514"),
            browser=self.browser,
        )

        await agent.run()

        # Parse the result and convert to Comment objects
        comments: list[Comment] = []

        # Note: This is a simplified implementation
        # In production, we would parse the actual response from the agent
        # For now, return empty list as placeholder
        logger.warning("Comment extraction needs implementation refinement")

        return comments

    async def post_reply(self, comment_id: str, text: str) -> bool:
        """Post a reply to a YouTube comment.

        Args:
            comment_id: YouTube comment ID
            text: Reply text

        Returns:
            True if successful, False otherwise
        """
        logger.info("Posting reply to comment %s", comment_id)

        agent = Agent(  # type: ignore[var-annotated]
            task=(
                f"Find the comment with ID {comment_id} and post a reply with the text: '{text}'. "
                f"Click the reply button, type the text, and submit."
            ),
            llm=ChatAnthropic(model="claude-sonnet-4-20250514"),
            browser=self.browser,
        )

        try:
            await agent.run()
        except Exception:
            logger.exception("Failed to post reply")
            return False
        else:
            logger.info("Reply posted successfully")
            return True
